==> CODE/6. Testing the M-TSP.ipynb <==
We start by creating our sample M-TSP problems, which we want to do in the following varities.

* N = 20, SP=3, SDP=2, EDP=2
* N = 40, SP=3, SDP=2, EDP=2
* N = 60, SP=3, SDP=2, EDP=2

The real problems range in size between N=40 and N=80. In addition, we want to solve each of these with the following time limits and compare results after. Let's generate the sample code for these three problems.

import random
import numpy as np
import json
from solver import *

def generate_random_coordinates():
    min_latitude = 51.855916482039945
    max_latitude = 51.9940501959335
    min_longitude = 3.9423262128546948
    max_longitude = 4.6041674797859695

    latitude = random.uniform(min_latitude, max_latitude)
    longitude = random.uniform(min_longitude, max_longitude)

    return([latitude, longitude])


generate_random_coordinates()

# Basis 
optimization_input = {
    'N20':{},
    'N40':{},
    'N60':{},
}

# Set seed
np.random.seed(100)

# Fill it 
for i in optimization_input:
    optimization_input[i] = {'WB_percentage':0.30,
            'SupplyBoats':{
                'Watergeus':103, # this refers to water capacity
                'Waterbuffel':103,
                'Waterval':83,
                'Waterman':83
            },
            'Depots':{
                'Starting':{
                    'SDP1':{
                        'Loc':[51.96249648967369, 4.137689887873459],
                        'Cap':1,
                    },
                    'SDP2':{
                        'Loc':[51.89740273617377, 4.460817447562738],
                        'Cap':6
                    }
                },
            'Ending':{
                    'EDP1':{
                    'Loc':[51.96249648967369, 4.137689887873459],
                    'Cap':1,
                },
                'EDP2':{
                    'Loc':[51.89740273617377, 4.460817447562738],
                    'Cap':6
                }
            }

            },
            'SupplyPoints':{
                'SP1':[51.96249648967369, 4.137689887873459],
                'SP2':[51.889365439900516, 4.313357332833194],
                'SP3':[51.89765499630826, 4.459774551250766]
            },
            'Shipments':{},
        }
    
optimization_input['N20']['Shipments'] = {'S{i}'.format(i=i+1):{'Loc':generate_random_coordinates(),'WaterAmount':random.randint(0, 30)} for i in range(20)}
optimization_input['N40']['Shipments'] = {'S{i}'.format(i=i+1):{'Loc':generate_random_coordinates(),'WaterAmount':random.randint(0, 30)} for i in range(40)}
optimization_input['N60']['Shipments'] = {'S{i}'.format(i=i+1):{'Loc':generate_random_coordinates(),'WaterAmount':random.randint(0, 30)} for i in range(60)}

Now, let's optimize these three test cases. We start with a time limit of 60 minutes. 

# Solve N20
id = 'N20'
sol = OptimizeRoutes(optimization_input[id], 3600)
    
# Store solution
solution = {
    'OBJ': sol[0],
    'Solving_time':sol[1],
    'Routes':sol[2],
    'Flow':sol[3],
    'Ordered_routes':sol[4],
    'Ordered_flow':sol[5]
    }

# Create json file
with open('solutions/testing/solution_{a}.json'.format(a=id), "w") as json_file:
    json.dump(solution, json_file)

# Solve N40
id = 'N40'
sol = OptimizeRoutes(optimization_input[id], 3600)
    
# Store solution
solution = {
    'OBJ': sol[0],
    'Solving_time':sol[1],
    'Routes':sol[2],
    'Flow':sol[3],
    'Ordered_routes':sol[4],
    'Ordered_flow':sol[5]
    }

# Create json file
with open('solutions/testing/solution_{a}.json'.format(a=id), "w") as json_file:
    json.dump(solution, json_file)

# Solve N60
id = 'N60'
sol = OptimizeRoutes(optimization_input[id], 3600*24)
    
# Store solution
solution = {
    'OBJ': sol[0],
    'Solving_time':sol[1],
    'Routes':sol[2],
    'Flow':sol[3],
    'Ordered_routes':sol[4],
    'Ordered_flow':sol[5]
    }

# Create json file
with open('solutions/testing/solution_{a}.json'.format(a=id), "w") as json_file:
    json.dump(solution, json_file)

import matplotlib.pyplot as plt
import numpy as np

# X-axis: time in minutes, from t=0 until t=60.
time = np.linspace(0, 60, num=61)

# Create random Y values for each of the 4 lines
tbl = pd.read_excel('extraction.xlsx.xlsx') # I've tracked these values manually in a XLSX file
N20_Incumbent = [x for x in tbl.loc[0]]
N20_BestBound = [x for x in tbl.loc[2]]

# Create the plot
fig, ax = plt.subplots(figsize=(10, 6))

# Add the 4 lines
ax.plot(time, N20_Incumbent, label='N20 - Incumbent', color='darkblue')
ax.plot(time, N20_BestBound, label='N20 - Best bound', color='crimson')
#ax.plot(time, N40_Incumbent, label='N40')

# Add X and Y axis labels
ax.set_xlabel('Time (minutes)', fontsize=12)
ax.set_ylabel('Objective value (distance in km)', fontsize=12)

# Add title
ax.set_title('Best objective value found over time (N=20, L=28)\n Final OBJ = 155.60 km, final gap = 21.11%', fontsize=15)

# Adjust the plot size to make room for the legend
fig.subplots_adjust(bottom=0.2)

# Add a legend below the plot
legend = ax.legend(loc='upper center', bbox_to_anchor=(0.5, -0.15), fancybox=True, shadow=True, ncol=4)

# Ensure the plot is nice and presentable
ax.grid(True)

# Save the plot
plt.savefig('plots/obj20.png', dpi=400)

# X-axis: time in minutes, from t=0 until t=60.
time = np.linspace(0, 60, num=61)

# Create random Y values for each of the 4 lines
tbl = pd.read_excel('extraction.xlsx.xlsx')
N40_Incumbent = [x for x in tbl.loc[1]]
N40_BestBound = [x for x in tbl.loc[3]]

# Create the plot
fig, ax = plt.subplots(figsize=(10, 6))

# Add the 4 lines
ax.plot(time, N40_Incumbent, label='N40 - Incumbent', color='darkblue')
ax.plot(time, N40_BestBound, label='N40 - Best bound', color='crimson')

# Add X and Y axis labels
ax.set_xlabel('Time (minutes)', fontsize=12)
ax.set_ylabel('Objective value (distance in km)', fontsize=12)

# Add title
ax.set_title('Best objective value found over time (N=40)\n Final OBJ = 282.83 km, final gap = 33.70%', fontsize=15)

# Adjust the plot size to make room for the legend
fig.subplots_adjust(bottom=0.2)

# Add a legend below the plot
legend = ax.legend(loc='upper center', bbox_to_anchor=(0.5, -0.15), fancybox=True, shadow=True, ncol=4)

# Ensure the plot is nice and presentable
ax.grid(True)

# Save the plot
plt.savefig('plots/obj40.png', dpi=400)

# X-axis: time in minutes, from t=0 until t=24*60 = 1440.
time = np.linspace(0, 24, num=24*6+1)

# Create random Y values for each of the 4 lines
tbl = pd.read_excel('extraction2.xlsx')
N60_Incumbent = [x for x in tbl.loc[1]]
N60_BestBound = [x for x in tbl.loc[2]]

# Create the plot
fig, ax = plt.subplots(figsize=(10, 6))

# Add the 4 lines
ax.plot(time, N60_Incumbent, label='N60 - Incumbent', color='darkblue')
ax.plot(time, N60_BestBound, label='N60 - Best bound', color='crimson')

# Add X and Y axis labels
ax.set_xlabel('Time (hours)', fontsize=12)
ax.set_ylabel('Objective value (distance in km)', fontsize=12)

# Add title
ax.set_title('Best objective value found over time (N=60, L=68)\n Final OBJ = 328.49 km, final gap = 32.11%', fontsize=15)

# Adjust the plot size to make room for the legend
fig.subplots_adjust(bottom=0.2)

# Add a legend below the plot
legend = ax.legend(loc='upper center', bbox_to_anchor=(0.5, -0.15), fancybox=True, shadow=True, ncol=4)

# Ensure the plot is nice and presentable
ax.grid(True)

# Save the plot
plt.savefig('plots/obj60.png', dpi=400)
