==> CODE/7. Extracting the optimization input.ipynb <==
Open the dataset

import pandas as pd
import json
evides = pd.read_csv('data/evides_withAIS.csv')

At the request of the PoR, shipments made to passenger ships are excluded from this analysis.

evides = evides[evides['Scheepstype'] != 'Passagiersschip, opleidingschip']

Extract opt_input

opt_input = {}

for i in set(evides.Datum):
    # Select subset for that date
    subset = evides[evides['Datum']==i]

    # Create dict for that date
    opt_input[i] = {'Shipments':{}}

    # Iterative over all shipments within subset
    s=1
    for j in subset.index:
        opt_input[i]['Shipments']['S{index}'.format(index=s)] = {'Loc':[evides['Latitude'].loc[j], evides['Longitude'].loc[j]], 'WaterAmount':evides['Hoeveelheid (m3)'].loc[j]}
        s=s+1

    # Take note of the amount of shipments
    opt_input[i]['N_Shipments'] = len(subset)

# Create json file
with open('data/optimization_input_regular.json', "w") as json_file:
    json.dump(opt_input, json_file)

opt_input

Show distribution

import matplotlib.pyplot as plt

# Create hist data
hist_data = [opt_input[x]['N_Shipments'] for x in opt_input]

# Plot hist data
fig, ax = plt.subplots(figsize=(12, 10),dpi=100)

ax.hist(hist_data, ec="black", color="lightseagreen", bins=40) 
ax.set_title("Distribution of the amount of shipments per day")
ax.set_xlabel("Number of shipments")
ax.set_ylabel("# instances")
ax.grid = True

# Then save a high-quality version of this image
plt.savefig('plots/distrshipments.png', dpi=400)

Repeat the same steps for the dataset in which we simulate a higher shipment limit.

# Import dataset
evides = pd.read_csv('data/evides_higherlimit.csv')
evides = evides[evides['Scheepstype'] != 'Passagiersschip, opleidingschip']

# Extract input
opt_input = {}

for i in set(evides.Datum):
    # Select subset for that date
    subset = evides[evides['Datum']==i]

    # Create dict for that date
    opt_input[i] = {'Shipments':{}}

    # Iterative over all shipments within subset
    s=1
    for j in subset.index:
        opt_input[i]['Shipments']['S{index}'.format(index=s)] = {'Loc':[evides['Latitude'].loc[j], evides['Longitude'].loc[j]], 'WaterAmount':evides['Hoeveelheid (m3)'].loc[j]}
        s=s+1

    # Take note of the amount of shipments
    opt_input[i]['N_Shipments'] = len(subset)

# Create json file
with open('data/optimization_input_highershipment.json', "w") as json_file:
    json.dump(opt_input, json_file)
