==> CODE/8. Running the main analysis.ipynb <==
In this notebook, we'll run the main analysis. For the sake of having an overview and being able to rerun each day inividually, I am running each simulation in its own cell block. Of course, this is not necessary at all. This notebook is a bit messier than the others.

# Import opt_input #1 - this is our input file for the regular evides data
import json
from solver import *
shipments = json.load(open('data/optimization_input_regular.json'))

Below, I'll identify the optimal routes for 10 instances of the problem. I'll initially allocate 4 hours to each problem.

def SolveOneInstance(id, time_limit, parameters, shipments):
    try:
        opt_input = parameters
        opt_input['Shipments'] = shipments[id]['Shipments']

        print(opt_input)
        
        sol = OptimizeRoutes(opt_input, time_limit)
            
        # Store solution
        solution = {
            'OBJ': sol[0],
            'Solving_time':sol[1],
            'Routes':sol[2],
            'Flow':sol[3],
            'Ordered_routes':sol[4],
            'Ordered_flow':sol[5],
            'Ordered_AIS':sol[6]
            }

        # Create json file
        with open('solutions/main_analysis/solution_{a}_regular.json'.format(a=id), "w") as json_file:
            json.dump(solution, json_file)
    
        return(print("Successfully found a solution for problem", id, "within the given timefrime."))
    
    except:
        return(print("The model was unable to find a solution for problem", id, "within the given timeframe."))

Denote fixed parameters

fixed_parameters = {'WB_percentage':0.30,
            'SupplyBoats':{
                'Watergeus':103, # this refers to water capacity
                'Waterbuffel':103,
                'Waterval':83,
                'Waterman':83
            },
            'Depots':{
                'Starting':{
                    'SDP1':{
                        'Loc':[51.96249648967369, 4.137689887873459],
                        'Cap':1,
                    },
                    'SDP2':{
                        'Loc':[51.89740273617377, 4.460817447562738],
                        'Cap':6
                    }
                },
            'Ending':{
                    'EDP1':{
                    'Loc':[51.96249648967369, 4.137689887873459],
                    'Cap':1,
                },
                'EDP2':{
                    'Loc':[51.89740273617377, 4.460817447562738],
                    'Cap':6
                }
            }

            },
            'SupplyPoints':{
                'SP1':[51.96249648967369, 4.137689887873459],
                'SP2':[51.889365439900516, 4.313357332833194],
                'SP3':[51.89765499630826, 4.459774551250766]
            },
            'Shipments':{},
        }

SolveOneInstance('2022-09-08', 4*3600, fixed_parameters, shipments)

SolveOneInstance('2022-09-29', 4*3600, fixed_parameters, shipments)

SolveOneInstance('2022-10-21', 4*3600, fixed_parameters, shipments)

SolveOneInstance('2022-01-27', 4*3600, fixed_parameters, shipments)

SolveOneInstance('2022-07-28', 4*3600, fixed_parameters, shipments)

SolveOneInstance('2022-07-06', 4*3600, fixed_parameters, shipments)

SolveOneInstance('2022-06-29', 4*3600, fixed_parameters, shipments)

SolveOneInstance('2022-10-14', 4*3600, fixed_parameters, shipments)

SolveOneInstance('2022-09-30', 4*3600, fixed_parameters, shipments)

SolveOneInstance('2022-03-09', 4*3600, fixed_parameters, shipments)

Let's also try to solve some instances with bigger problem sizes!

SolveOneInstance('2022-05-03', 6*3600, fixed_parameters, shipments) # N=60

SolveOneInstance('2022-10-04', 6*3600, fixed_parameters, shipments) # N=60

SolveOneInstance('2022-05-19', 6*3600, fixed_parameters, shipments) # N=60

SolveOneInstance('2022-03-08', 12*3600, fixed_parameters, shipments) # N=70

SolveOneInstance('2022-07-26', 12*3600, fixed_parameters, shipments) # N=70

Now we re-run the same code, using the optimization_input JSON generated with the higher shipment dataset.

# Import other JSON
shipments = json.load(open('data/optimization_input_highershipment.json'))
shipments

# Redefine function
def SolveOneInstance(id, time_limit, parameters, shipments):
    try:
        opt_input = parameters
        opt_input['Shipments'] = shipments[id]['Shipments']

        print(opt_input)
        
        sol = OptimizeRoutes(opt_input, time_limit)
            
        # Store solution
        solution = {
            'OBJ': sol[0],
            'Solving_time':sol[1],
            'Routes':sol[2],
            'Flow':sol[3],
            'Ordered_routes':sol[4],
            'Ordered_flow':sol[5],
            'Ordered_AIS':sol[6]
            }

        # Create json file
        with open('solutions/main_analysis/solution_{a}_higherlimit.json'.format(a=id), "w") as json_file:
            json.dump(solution, json_file)
    
        return(print("Successfully found a solution for problem", id, "within the given timefrime."))
    
    except:
        return(print("The model was unable to find a solution for problem", id, "within the given timeframe."))

shipments

SolveOneInstance('2022-09-08', 4*3600, fixed_parameters, shipments)

SolveOneInstance('2022-09-29', 4*3600, fixed_parameters, shipments)

SolveOneInstance('2022-10-21', 4*3600, fixed_parameters, shipments)

SolveOneInstance('2022-01-27', 4*3600, fixed_parameters, shipments)

SolveOneInstance('2022-07-28', 4*3600, fixed_parameters, shipments)

SolveOneInstance('2022-07-06', 4*3600, fixed_parameters, shipments)

SolveOneInstance('2022-06-29', 4*3600, fixed_parameters, shipments)

SolveOneInstance('2022-10-14', 4*3600, fixed_parameters, shipments)

SolveOneInstance('2022-09-30', 4*3600, fixed_parameters, shipments)

SolveOneInstance('2022-03-09', 4*3600, fixed_parameters, shipments)

And we do the same for the larger shipments.

SolveOneInstance('2022-05-03', 6*3600, fixed_parameters, shipments) 

SolveOneInstance('2022-10-04', 6*3600, fixed_parameters, shipments)

SolveOneInstance('2022-03-08', 12*3600, fixed_parameters, shipments)

SolveOneInstance('2022-07-26', 12*3600, fixed_parameters, shipments)

Since after running for the first time there still appear to be shipments included in the opt_input which are better left out (i.e. those in Dordrecht, or even Nijmegen!), I'll run some extra code to get rid of these. Note that this is not necessary if the process in notebook 1. AIS extraction is improved or made redundant. 

# Redefine function
def SolveOneInstance(id, time_limit, parameters, shipments, savename):
    try:
        opt_input = parameters
        opt_input['Shipments'] = shipments[id]['Shipments']

        print(opt_input)
        
        sol = OptimizeRoutes(opt_input, time_limit)
            
        # Store solution
        solution = {
            'OBJ': sol[0],
            'Solving_time':sol[1],
            'Routes':sol[2],
            'Flow':sol[3],
            'Ordered_routes':sol[4],
            'Ordered_flow':sol[5],
            'Ordered_AIS':sol[6]
            }

        # Create json file
        with open('solutions/main_analysis/solution_{a}_{savename}.json'.format(a=id, savename=savename), "w") as json_file:
            json.dump(solution, json_file)
    
        return(print("Successfully found a solution for problem", id, "within the given timefrime."))
    
    except:
        return(print("The model was unable to find a solution for problem", id, "within the given timeframe."))

# shipments (regular)
shipments = json.load(open('data/optimization_input_regular.json'))

def location_in_area(location, area):
    (px, py) = location
    ((rx_min, ry_min), (rx_max, ry_max)) = area

    # Check if within boundaries
    if rx_min <= px <= rx_max and ry_min <= py <= ry_max:
        return 1
    else:
        return 0

for x in shipments:
    a = 0
    to_delete = []
    for y in shipments[x]['Shipments']:
        rm = location_in_area(shipments[x]['Shipments'][y]['Loc'], ((51.86, 3.9), (52, 4.6)))
        if rm == 0:
            to_delete.append(y)
    
    for key in to_delete:
        del shipments[x]['Shipments'][key]

    # rename shipments
    counter = 1
    new_shipments = {}
    for y in shipments[x]['Shipments']:
        new_shipments['S' + str(counter)] = shipments[x]['Shipments'][y]
        counter += 1

    shipments[x]['Shipments'] = new_shipments
    shipments[x]['N_Shipments'] = len(new_shipments)

shipments

Then let's repeat our calculations

# Solve instances (regular)
SolveOneInstance('2022-01-27', 4*3600, fixed_parameters, shipments, 'regular')
SolveOneInstance('2022-03-08', 12*3600, fixed_parameters, shipments, 'regular')
SolveOneInstance('2022-03-09', 4*3600, fixed_parameters, shipments, 'regular')
SolveOneInstance('2022-05-19', 6*3600, fixed_parameters, shipments, 'regular')
SolveOneInstance('2022-09-29', 4*3600, fixed_parameters, shipments, 'regular')
SolveOneInstance('2022-10-04', 6*3600, fixed_parameters, shipments, 'regular')
SolveOneInstance('2022-10-14', 12*3600, fixed_parameters, shipments, 'regular')

Then we do the same for our higher shipment set

# shipments (regular)
shipments = json.load(open('data/optimization_input_highershipment.json'))

for x in shipments:
    a = 0
    to_delete = []
    for y in shipments[x]['Shipments']:
        rm = location_in_area(shipments[x]['Shipments'][y]['Loc'], ((51.86, 3.9), (52, 4.6)))
        if rm == 0:
            to_delete.append(y)
    
    for key in to_delete:
        del shipments[x]['Shipments'][key]

    # rename shipments
    counter = 1
    new_shipments = {}
    for y in shipments[x]['Shipments']:
        new_shipments['S' + str(counter)] = shipments[x]['Shipments'][y]
        counter += 1

    shipments[x]['Shipments'] = new_shipments
    shipments[x]['N_Shipments'] = len(new_shipments)

shipments

# Solve instances (higher limit)
SolveOneInstance('2022-01-27', 4*3600, fixed_parameters, shipments, 'higherlimit')
SolveOneInstance('2022-03-08', 12*3600, fixed_parameters, shipments, 'higherlimit')
SolveOneInstance('2022-03-09', 4*3600, fixed_parameters, shipments, 'higherlimit')
SolveOneInstance('2022-05-19', 6*3600, fixed_parameters, shipments, 'higherlimit')
SolveOneInstance('2022-09-29', 4*3600, fixed_parameters, shipments, 'higherlimit')
SolveOneInstance('2022-10-04', 6*3600, fixed_parameters, shipments, 'higherlimit')
SolveOneInstance('2022-10-14', 12*3600, fixed_parameters, shipments, 'higherlimit')
